update cuyy, delete jadwal masih ngebug

+++ include/studio.h
@@ -2,8 +2,8 @@
 #define STUDIO_H
 
 #define MAX_STUDIO_NAME 100
-#define STUDIO_SETTER_FORMAT "%d,%s,%d,%d,%f\n"
-#define STUDIO_GETTER_FORMAT "%d,%[^,],%d,%d,%f\n"
+#define STUDIO_SETTER_FORMAT "%d,%s,%d,%d,%d\n"
+#define STUDIO_GETTER_FORMAT "%d,%[^,],%d,%d,%d\n"
 #define STUDIO_DATABASE_NAME "db/studio.csv"
 
 // ================================== type definition ================================== //
@@ -14,7 +14,7 @@ typedef struct
     char nama[MAX_STUDIO_NAME];
     int bioskop_id; // Foreign key ke tabel Bioskop
     int jumlah_kursi;
-    float additionalFee;
+    int additionalFee;
 } Studio;
 
 // ================================== setter ================================== //
@@ -22,14 +22,14 @@ typedef struct
 void setNamaStudio(Studio *studio, const char *nama);
 void setStudioBioskopId(Studio *studio, int bioskop_id);
 void setJumlahKursi(Studio *studio, int jumlah_kursi);
-void setHarga(Studio *studio, float additionalFee);
+void setHarga(Studio *studio, int additionalFee);
 
 // ================================== getter ================================== //
 
 const char *getNamaStudio(const Studio *studio);
 int getStudioBioskopId(const Studio *studio);
 int getJumlahKursi(const Studio *studio);
-float getHarga(const Studio *studio);
+int getHarga(const Studio *studio);
 
 // ================================== Main Menu ================================ //
 
@@ -40,8 +40,8 @@ void updateStudioMenu(Studio studio);
 // ================================== Action ================================== //
 
 Studio *findStudioById(int id);
-Studio *createStudio(const char *nama, int jumlah_kursi, float additionalFee);
-Studio *updateStudio(const int id, const char *nama, int jumlah_kursi, float additionalFee);
+Studio *createStudio(const char *nama, int jumlah_kursi, int additionalFee);
+Studio *updateStudio(const int id, const char *nama, int jumlah_kursi, int additionalFee);
 int deleteStudio(Studio studio);
 
 // ================================== Utils ================================== //
